701 refactor editor tab into its own plugin (#711)
* Moved MainScreenshotPlugin from a separate plugin into a service for the EditorTab plugin

Various refactoring

* Migrated drag+drop into the main plugin

* Lots of removal of SelectionManager from Gum

* More SelectionManager refactoring

* Killed SelectionManager singleton

* Removed EditingManager Singleton

* More refactoring

* Removed GuiCommands old

* Made WireframeObjectManager have class scoped namespace

* Fixed negative font file

* fixed scrollbar bug

* More cleanup of WireframeControl

* More cleanup of WireframeObjectManager

* Another cleanup pass

* Small cleanup

* Initial migration of editor tab to its own plugin

* Editor Tab Plugin is now its own separate plugin
