+++ Gum.sln
@@ -53,6 +53,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GumCommon", "GumCommon\GumC
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonFormsAndControls", "CommonFormsAndControls\CommonFormsAndControls.csproj", "{B7ABE8DF-EE36-42AB-B371-78195C614A61}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "EditorTabPlugins", "EditorTabPlugins", "{3E451941-5A67-42CD-931E-5DB8C0666178}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EditorTabPlugin_XNA", "Tool\EditorTabPlugin_XNA\EditorTabPlugin_XNA.csproj", "{F4A5BB2F-1650-4099-9FFE-1C4A416D4438}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x86 = Debug|x86
@@ -135,6 +139,10 @@ Global
 		{B7ABE8DF-EE36-42AB-B371-78195C614A61}.Debug|x86.Build.0 = Debug|x86
 		{B7ABE8DF-EE36-42AB-B371-78195C614A61}.Release|x86.ActiveCfg = Release|x86
 		{B7ABE8DF-EE36-42AB-B371-78195C614A61}.Release|x86.Build.0 = Release|x86
+		{F4A5BB2F-1650-4099-9FFE-1C4A416D4438}.Debug|x86.ActiveCfg = Debug|x86
+		{F4A5BB2F-1650-4099-9FFE-1C4A416D4438}.Debug|x86.Build.0 = Debug|x86
+		{F4A5BB2F-1650-4099-9FFE-1C4A416D4438}.Release|x86.ActiveCfg = Release|x86
+		{F4A5BB2F-1650-4099-9FFE-1C4A416D4438}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -161,6 +169,8 @@ Global
 		{29DBC0A6-F95A-442A-A063-B722DDCB02C6} = {C66AD07F-C683-4C24-8E6B-20638ACC3A25}
 		{6F20DA16-FDC2-485E-B6ED-433EE2A97E59} = {C66AD07F-C683-4C24-8E6B-20638ACC3A25}
 		{B7ABE8DF-EE36-42AB-B371-78195C614A61} = {C66AD07F-C683-4C24-8E6B-20638ACC3A25}
+		{3E451941-5A67-42CD-931E-5DB8C0666178} = {28BBCA70-AFC9-4C9B-9114-E0E23FA01C3F}
+		{F4A5BB2F-1650-4099-9FFE-1C4A416D4438} = {3E451941-5A67-42CD-931E-5DB8C0666178}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {691B2E2F-58CB-4FD2-8480-695E8DE0CCE2}
