+++ Tool/EditorTabPlugin_XNA/Editors/DimensionDisplay.cs
