+++ Tool/EditorTabPlugin_XNA/Editors/ResizeHandles.cs
