+++ Tool/EditorTabPlugin_XNA/Views/WireframeEditControl.resx
